backend/src/chatbot/requirements.txt
fastapi
langchain
langchain-chroma
langchain-core
langchain-community
langchain-huggingface
langchain-openai
llama-index
llama-parse
openai
pydantic
pytest
python-dotenv
requests
typing
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
uvicorn
